requirements.txt
pymupdf
chromadb
pydantic-ai
pydantic
openai
tiktoken
python-dotenv
